Migrate off of CROSSTOOL.
This sets things up to move to a newer version of Bazel.

Largely followed this tutorial:
https://docs.bazel.build/versions/0.24.0/tutorial/cc-toolchain-config.html

Had to tweak a couple of minor things in the code to get this to build,
probably thanks to stricter defaults.

+++ mosh_nacl/mosh_nacl.cc
@@ -720,7 +720,7 @@ int sigaction(int signum, const struct sigaction* act,
   return 0;
 }
 
-int ioctl(int d, long unsigned int request, ...) {  // NOLINT(runtime/int)
+int ioctl(int d, int request, ...) {  // NOLINT(runtime/int)
   if (d != STDIN_FILENO || request != TIOCGWINSZ) {
     Log("ioctl(%d, %u, ...): Got unexpected call", d, request);
     errno = EPROTO;
